CookieTest: split `path_matches()` specific tests to dedicated test class
